refactor(NODE-3717): test reorg part 4 (#3086)
