+++ test/integration/server-discovery-and-monitoring/server_discovery_and_monitoring.spec.test.js
@@ -1,7 +1,6 @@
 'use strict';
-const TestRunnerContext = require('./spec-runner').TestRunnerContext;
-const loadSpecTests = require('../spec').loadSpecTests;
-const generateTopologyTests = require('./spec-runner').generateTopologyTests;
+const { TestRunnerContext, generateTopologyTests } = require('../../tools/spec-runner');
+const { loadSpecTests } = require('../../spec');
 
 class SDAMRunnerContext extends TestRunnerContext {
   constructor() {
